pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiobotocore>=3.2.1",
    "alembic>=1.18.4",
    "annotated-types>=0.7.0",
    "anyio>=4.12.1",
    "asyncpg>=0.31.0",
    "attrs>=25.4.0",
    "babel>=2.18.0",
    "bcrypt>=5.0.0",
    "brotli>=1.2.0",
    "brotlicffi>=1.2.0.1",
    "certifi>=2026.2.25",
    "cffi>=2.0.0",
    "click>=8.3.1",
    "colorama>=0.4.6",
    "curio>=1.6",
    "dishka>=1.9.1",
    "dnspython>=2.8.0",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.115.0",
    "h11>=0.16.0",
    "h2>=4.3.0",
    "hpack>=4.1.0",
    "httpcore[asyncio,http2,socks,trio]>=1.0.9",
    "httptools>=0.7.1",
    "httpx[brotli,http2,socks,zstd]>=0.28.1",
    "hyperframe>=6.1.0",
    "idna>=3.11",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "markdown-it-py>=4.0.0",
    "markupsafe>=3.0.3",
    "mdurl>=0.1.2",
    "outcome>=1.3.0.post0",
    "pwdlib[argon2]>=0.3.0",
    "pillow>=11.0.0",
    "pycparser>=3.0",
    "pygments>=2.19.2",
    "pyjwt[crypto]>=2.12.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "pytz>=2026.1.post1",
    "pyyaml>=6.0.3",
    "redis-om>=1.0.4b0",
    "redis[hiredis]>=7.3.0",
    "rich>=14.3.3",
    "setuptools>=82.0.0",
    "sniffio>=1.3.1",
    "socksio>=1.0.0",
    "sortedcontainers>=2.4.0",
    "sqlalchemy[asyncio]>=2.1.0b1",
    "starlette[full]>=0.52.1",
    "structlog>=25.5.0",
    "taskiq>=0.12.1",
    "taskiq-aio-pika>=0.6.0",
    "trio>=0.33.0",
    "typing-extensions>=4.15.0",
    "tzdata>=2025.3",
    "uvicorn[standard]>=0.41.0",
    "watchfiles>=1.1.1",
    "websockets>=16.0",
    "zstandard>=0.25.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "locust>=2.43.3",
    "mypy>=1.19.1",
    "polyfactory>=3.3.0",
    "pyclean>=3.5.0",
    "pytest>=9.0.2",
    "pytest-archon>=0.0.7",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "testcontainers[minio,postgres,rabbitmq,redis]>=4.14.1",
]
[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "architecture: Fitness functions and boundary enforcement",
    "unit: Domain-layer pure logic, zero I/O",
    "integration: Application + Infrastructure with real database",
    "e2e: Presentation-layer HTTP round-trips",
    "load: Resilience and threshold testing (Locust)",
]
filterwarnings = ["ignore::DeprecationWarning:dishka.*"]
testpaths = ["tests"]

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501", "RUF001", "RUF002", "RUF003", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.14"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
